+++ src/main/java/com/venus/modules/sys/controller/SysRoleController.java
@@ -2,6 +2,7 @@ package com.venus.modules.sys.controller;
 
 import com.venus.common.annotation.LogOperation;
 import com.venus.common.constant.Constant;
+import com.venus.common.exception.ErrorCode;
 import com.venus.common.page.PageData;
 import com.venus.common.utils.Result;
 import com.venus.common.validator.AssertUtils;
@@ -10,9 +11,10 @@ import com.venus.common.validator.group.AddGroup;
 import com.venus.common.validator.group.DefaultGroup;
 import com.venus.common.validator.group.UpdateGroup;
 import com.venus.modules.sys.dto.SysRoleDTO;
-import com.venus.modules.sys.service.SysRoleDataScopeService;
-import com.venus.modules.sys.service.SysRoleMenuService;
-import com.venus.modules.sys.service.SysRoleService;
+import com.venus.modules.sys.dto.SysRoleUsersDTO;
+import com.venus.modules.sys.dto.SysUserDTO;
+import com.venus.modules.sys.entity.SysRoleUserEntity;
+import com.venus.modules.sys.service.*;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
@@ -22,6 +24,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import springfox.documentation.annotations.ApiIgnore;
 
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -35,7 +38,11 @@ public class SysRoleController {
     @Autowired
     private SysRoleMenuService sysRoleMenuService;
     @Autowired
+    private SysRoleUserService sysRoleUserService;
+    @Autowired
     private SysRoleDataScopeService sysRoleDataScopeService;
+    @Autowired
+    private SysUserService sysUserService;
 
     @GetMapping("page")
     @ApiOperation("分页")
@@ -99,7 +106,7 @@ public class SysRoleController {
         return new Result();
     }
 
-    @DeleteMapping
+    @DeleteMapping("delete")
     @ApiOperation("删除")
     @LogOperation("删除")
     @RequiresPermissions("sys:role:delete")
@@ -108,7 +115,52 @@ public class SysRoleController {
         //效验数据
         AssertUtils.isArrayEmpty(ids, "id");
 
+        Long count = sysRoleUserService.countByRoleIds(ids);
+
+        if(count > 0) {
+            return new Result().error(ErrorCode.ROLE_USER_EXIST);
+        }
+
         sysRoleService.delete(ids);
+
+        return new Result();
+    }
+
+    @GetMapping("users")
+    @ApiOperation("用户列表")
+    @LogOperation("获取用户列表")
+    public Result<PageData<SysUserDTO>> users(@ApiIgnore @RequestParam Map<String, Object> params) {
+        String roleIdStr = (String) params.get("roleId");
+
+        AssertUtils.isNull(roleIdStr, "roleId");
+        params.put("roleId", Long.valueOf(roleIdStr));
+        PageData<SysUserDTO> page = sysUserService.getListByRoleId(params);
+
+        return new Result<PageData<SysUserDTO>>().ok(page);
+    }
+
+    @PostMapping("saveUsers")
+    @ApiOperation("保存用户")
+    @LogOperation("保存用户")
+    @RequiresPermissions("sys:role:save")
+    public Result saveUsers(@RequestBody SysRoleUsersDTO sysRoleUsersDTO) {
+        // 效验数据
+        ValidatorUtils.validateEntity(sysRoleUsersDTO, AddGroup.class, DefaultGroup.class);
+
+        sysRoleUserService.saveByRoleId(sysRoleUsersDTO.getRoleId(), sysRoleUsersDTO.getUserIds());
+
+        return new Result();
+    }
+
+    @DeleteMapping("deleteUsers")
+    @ApiOperation("删除用户")
+    @LogOperation("删除用户")
+    @RequiresPermissions("sys:role:delete")
+    public Result deleteUsers(@RequestBody SysRoleUsersDTO sysRoleUsersDTO) {
+        // 效验数据
+        ValidatorUtils.validateEntity(sysRoleUsersDTO, UpdateGroup.class, DefaultGroup.class);
+
+        sysRoleUserService.deleteByRoleId(sysRoleUsersDTO.getRoleId(), sysRoleUsersDTO.getUserIds());
 
         return new Result();
     }
